Lots of changes: see the description for details
added the alertInvalFileFormat method which lets the user know that the
first 4 bytes of file are invalid.  This then creates a completely blank
gameboard.  I had to add back the constructors in SideButtons and
GridButtons that set up a blank game.  Also had to change the setup
function in GameWindow so that it know accepts a boolean blank which
tells it whether or not to set up a game with a maze or a blank one.