Added println statements in Main for testing.......
The default.mze file in Program 6 is a played game according to println
statement.  I asked Dr. Buckner if we can have an original game for
testing purposes.  Added the variable long time which is the time in
seconds.  There is a println so you can see the long  that is read.
Fixed a bug in ByteFileStreamReader I previously created.