Corrected finding from Program 5.  See description for detail.
Dr. Buckner found that if his current directory did not have a .mze file
and he tried to navigate to another directory, he could not successfully
load a .mze file.  The changes to the JFileChooser allow it to select
from both Files and Directories and then get the files absolute path.