Corrected a finding that Dr. Buckner pointed out to me yesterday
If a user tries to load a bad file before a gameboard is ever created,
there is a pop up message that allows them to open a file or cancel.
Dr. Buckner selected cancel and was ejected from the game.  He did not
like this.  So, I created a boolean called gameBoardIsPresent.  If a
gameboard was previously created, then it is true and when the user
selects cancel, it defaults to the board they were using.  If a
gameboard never existed, then it gives them a pop up message that a
blank game will be created and then a blank game comes up and the
boolean is reset to true.