Fixed bug in load() related to Buckner's finding in Program 5: see description below
Because I had to duplicate code to create the windows shell for the game
in the load method in GameWindow, I created a new function in GameWindow
called createWindowShell().  It loads in the dimensions of the window,
default colors, and default behavior on exit.  This function replaces
those lines we had in Main with a call to game.createWindowShell()